feature: adicionado wiremock para mockar servico externo
