+++ docker-compose.yml
@@ -1,10 +1,4 @@
 services:
-  wiremock:
-    image: "wiremock/wiremock:latest"
-    container_name: wiremock-standalone
-    entrypoint: [ "/docker-entrypoint.sh", "--global-response-templating", "--disable-gzip", "--verbose" ]
-    ports:
-      - "8080:8080"
 
   postgres:
     image: postgres:15
@@ -20,4 +14,3 @@ services:
 
 volumes:
   postgres_data:
-  wiremock_data:
