+++ src/main/resources/application-test.yml
@@ -1,5 +1,5 @@
-authorization-dev-tools:
-  url: http://localhost:8080
-
 server:
   port: 9090
+
+authorization-client:
+  base-url: http://localhost:9000
