+++ src/main/resources/application.properties
@@ -15,3 +15,5 @@ spring.flyway.baseline-on-migrate=true
 logging.level.root=INFO
 
 server.port=8081
+
+authorization-client.base-url=https://util.devi.tools/api/v2
