+++ src/test/java/SimplifyPay/infrastructure/TestConfig.java
@@ -1,6 +1,5 @@
 package SimplifyPay.infrastructure;
 
-import com.github.tomakehurst.wiremock.WireMockServer;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 
@@ -15,10 +14,4 @@ public class TestConfig {
     public UserTestScenario userTestScenario() {
         return new UserTestScenario();
     }
-
-    @Bean
-    public WireMockServer wireMockServer() {
-        return new WireMockServer(8080);
-    }
-
 }
